ui: Make use of ImPlot to drastically improve information view

+++ include/views/view_information.hpp
@@ -26,7 +26,7 @@ namespace hex {
         float m_highestBlockEntropy = 0;
         std::vector<float> m_blockEntropy;
 
-        std::array<float, 256> m_valueCounts = { 0 };
+        std::array<u64, 256> m_valueCounts = { 0 };
         bool m_analyzing = false;
 
         std::pair<u64, u64> m_analyzedRegion = { 0, 0 };
